Move team routes to the settings file and delete files related to applications/team
